Fix db table recreation

+++ src/database.py
@@ -36,7 +36,8 @@ async def get_db_session() -> AsyncGenerator[AsyncSession, None]:
             raise
 
 
-async def recreate_database(engine: AsyncEngine) -> None:
+async def create_db_tables(engine: AsyncEngine, drop: bool = False) -> None:
     async with engine.begin() as conn:  # `engine.begin()` due to a synchronous DB commands
-        await conn.run_sync(db.Base.metadata.drop_all)
+        if drop:
+            await conn.run_sync(db.Base.metadata.drop_all)
         await conn.run_sync(db.Base.metadata.create_all)
